# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Assinatura Eletrônica Avançada",
        "description": "API focada em manipulação de pdf com o foco em Assinatura virtual",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3333/v1",
            "description": "API DE TESTE"
        }
    ],
    "paths": {
        "/documents/{documentId}": {
            "get": {
                "summary": "Visualizar dados do documento",
                "description": "Essa rota ira buscar os dados do documento de acordo com o ID dele",
                "tags": ["Documents"],
                "parameters": [
                    {
                        "name": "documentId",
                        "in": "path",
                        "description": "Id para a busca do documento",
                        "required": true
                    }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "401": {
                        "description": "Não autorizado"
                    },
                    "400": {
                        "description": "Documento não encontrado"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/doc"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Deletar o documento",
                "description": "Essa rota ira deletar o documento atraves do Id",
                "tags": ["Documents"],
                "parameters": [
                    {
                        "name": "documentId",
                        "in": "path",
                        "description": "Id para a busca do documento",
                        "required": true
                    }
                ],
                "security": [{ "bearerAuth": [] }],
                "response": {
                    "500": {
                        "description": "erro description"
                    },
                    "400": {
                        "description": "Documento nao encontrado"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "200": {
                        "description": "Documento excluido com sucesso"
                    }
                }
            }
        },
        "/documents/{documentId}/prepare-signature": {
            "get": {
                "summary": "Selecionar o campo de assinatura",
                "description": "Essa rota ira selecionar a posição no documento que deverá ficar a assinatura",
                "tags": ["Signature"],
                "parameters": [
                    {
                        "name": "documentId",
                        "in": "path",
                        "description": "Id para a busca do documento",
                        "required": true
                    }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "401": {
                        "description": "Não autorizado"
                    },
                    "400": {
                        "description": "Documento não encontrado"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/html": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/documents/{documentId}/download": {
            "get": {
                "summary": "Baixar o PDF",
                "description": "Após assinado o documento estará disponivel para download",
                "tags": ["Documents", "Download"],
                "parameters": [
                    {
                        "name": "documentId",
                        "in": "path",
                        "description": "Id para a busca do documento",
                        "required": true
                    }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "401": {
                        "description": "Não autorizado"
                    },
                    "409": {
                        "description": "Documento não está completo"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/pdf": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/documents": {
            "post": {
                "summary": "Enviar o PDF",
                "description": "O PDF é enviado, dando inicio ao processo",
                "tags": ["Documents"],
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "content": {
                        "application/pdf": {
                            "schema": {
                                "type": "string",
                                "format": "binary"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Não autorizado"
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "documentId": {
                                            "type": "number",
                                            "format": "int"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    },
                                    "example": {
                                        "documentId": "1746582479887",
                                        "status": "uploaded"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/documents/{documentId}/signers": {
            "post": {
                "summary": "Adicionar Assinantes",
                "description": "Adicionar aqueles que devem assinar o documento",
                "tags": ["Signature"],
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    {
                        "name": "documentId",
                        "in": "path",
                        "description": "Id para a busca do documento",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/signer"
                            }
                        }
                    }
                },
                "result": {
                    "400": {
                        "description": "Erro ao adicionar signatário."
                    },
                    "200": {
                        "description": "Signatários adicionados com sucesso."
                    }
                }
            }
        },
        "/documents/{documentId}/signature-fields": {
            "post": {
                "summary": "Posicionar Campo",
                "description": "Posicionar campo de assinatura na posição correta",
                "tags": ["Signature", "Fields"],
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    {
                        "name": "documentId",
                        "in": "path",
                        "description": "Id para a busca do documento",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/field"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Não foi possivel adicionar o campo"
                    },
                    "200": {
                        "description": "Campo adicionado com sucesso"
                    }
                }
            }
        },
        "/documents/{documentId}/sign": {
            "post": {
                "summary": "Assinar Documento",
                "description": "O usuario adicionado no final '/signers' assina o documento",
                "tags": ["Signature"],
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    {
                        "name": "documentId",
                        "in": "path",
                        "description": "Id para a busca do documento",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/signer"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Documento Assinado."
                    },
                    "200": {
                        "description": "Erro ao assinar o documento"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "signer": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "order": {
                        "type": "number",
                        "format": "int"
                    }
                },
                "example": {
                    "name": "Fulano Beltrano",
                    "email": "user@example.com",
                    "order": 1
                }
            },
            "doc": {
                "type": "object",
                "properties": {
                    "documentId": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "signer": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "format": "email"
                                },
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "example": {
                    "documentId": "1746582479887",
                    "status": "in_progress",
                    "signer": [
                        {
                            "email": "user@example.com",
                            "status": "pending"
                        }
                    ]
                }
            },
            "field": {
                "type": "object",
                "properties": {
                    "x": {
                        "type": "number",
                        "format": "float"
                    },
                    "y": {
                        "type": "number",
                        "format": "float"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    }
                },
                "example": {
                    "x": 52.053,
                    "y": 89.009,
                    "email": "user@example.com"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}